refactor: move offline indicator to header menu
- Add offline indicator next to timetable link in Header/Menu
- Remove separate OfflineBanner component from timetable page
- Show indicator when backend is unreachable
